wrong place for the newQuoteSchema back to client
